requirements.txt
discord.py
aiohttp
PIL
io
pathlib